ai_backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
